requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
openai
google-generativeai